// src/mqtt-hardware/enums/hardware-commands.enum.ts
export enum HardwareCommand {
  LIGAR_BOMBA = 'LIGAR_BOMBA',
  DESLIGAR_BOMBA = 'DESLIGAR_BOMBA',
  ABRIR_VALVULA = 'ABRIR_VALVULA',
  FECHAR_VALVULA = 'FECHAR_VALVULA',
  INICIAR_PROCESSO_VACUO = 'INICIAR_PROCESSO_VACUO',
  PARAR_PROCESSO = 'PARAR_PROCESSO',
  PARADA_EMERGENCIA = 'PARADA_EMERGENCIA',
  DESLIGAR_TODAS_BOMBAS = 'DESLIGAR_TODAS_BOMBAS',
  DESLIGAR_TODAS_VALVULAS = 'DESLIGAR_TODAS_VALVULAS',
  SINCRONIZAR_HARDWARE = 'SINCRONIZAR_HARDWARE',
}